All chest bizz really...dayem..
Got the iron chests spawning loot in a much more pleasurable way.. tho the loot spawns from the middle of the chest when the lid is open.. when I should look like it's comming from inside the box.. how do I move the x or z values to fix?

Also the gold chest is being a pain in the ass.. I want to only have charms.. but if I don't have at lease one object in the object slot..then the chest will often open with nothing inside.  This object gets pulled 50% of the time... and the rest of the 30 charms are the other 50% (written up on trello)